+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- Pin table column to the left when grouping by a corresponding field is applied (roubkar)
 - Disable highlight mode control if metrics are aggregated (gorarakelyan)
 - Apply color, stroke and grid pattern repeating logic (roubkar)
 - Add ability to select mode for highlighting metric plots (roubkar)